docs: fix path issue
